requirements.txt
transformers
torch
streamlit
openai==0.28.1
python-dotenv